fix(serialization): sort of fix loading
